+++ engine/baml-runtime/src/runtime/runtime_interface.rs
@@ -295,8 +295,6 @@ impl InternalRuntimeInterface for InternalBamlRuntime {
         test_name: &str,
         ctx: &RuntimeContextManager,
     ) -> Result<Option<TypeBuilder>> {
-        use crate::type_builder::WithMeta;
-
         let func = self.get_function(function_name, &ctx.create_ctx(None, None)?)?;
         let test = self.ir().find_test(&func, test_name)?;
 
@@ -306,81 +304,7 @@ impl InternalRuntimeInterface for InternalBamlRuntime {
 
         let type_builder = TypeBuilder::new();
 
-        for entry in test.type_builder_contents() {
-            match entry {
-                TypeBuilderEntry::Class(cls) => {
-                    let mutex = type_builder.class(&cls.elem.name);
-                    let class_builder = mutex.lock().unwrap();
-                    for f in &cls.elem.static_fields {
-                        class_builder
-                            .property(&f.elem.name)
-                            .lock()
-                            .unwrap()
-                            .r#type(f.elem.r#type.elem.to_owned())
-                            .with_meta(
-                                "alias",
-                                f.attributes.get("alias").map_or(BamlValue::Null, |v| {
-                                    v.resolve_string(&EvaluationContext::default())
-                                        .map_or(BamlValue::Null, BamlValue::String)
-                                }),
-                            )
-                            .with_meta(
-                                "description",
-                                f.attributes
-                                    .get("description")
-                                    .map_or(BamlValue::Null, |v| {
-                                        v.resolve_string(&EvaluationContext::default())
-                                            .map_or(BamlValue::Null, BamlValue::String)
-                                    }),
-                            );
-                    }
-                }
-
-                TypeBuilderEntry::Enum(enm) => {
-                    let mutex = type_builder.r#enum(&enm.elem.name);
-                    let enum_builder = mutex.lock().unwrap();
-                    for (variant, _) in &enm.elem.values {
-                        enum_builder
-                            .value(&variant.elem.0)
-                            .lock()
-                            .unwrap()
-                            .with_meta(
-                                "alias",
-                                variant
-                                    .attributes
-                                    .get("alias")
-                                    .map_or(BamlValue::Null, |v| {
-                                        v.resolve_string(&EvaluationContext::default())
-                                            .map_or(BamlValue::Null, BamlValue::String)
-                                    }),
-                            )
-                            .with_meta(
-                                "description",
-                                variant.attributes.get("description").map_or(
-                                    BamlValue::Null,
-                                    |v| {
-                                        v.resolve_string(&EvaluationContext::default())
-                                            .map_or(BamlValue::Null, BamlValue::String)
-                                    },
-                                ),
-                            )
-                            .with_meta(
-                                "skip",
-                                variant.attributes.get("skip").map_or(BamlValue::Null, |v| {
-                                    v.resolve_bool(&EvaluationContext::default())
-                                        .map_or(BamlValue::Null, BamlValue::Bool)
-                                }),
-                            );
-                    }
-                }
-
-                TypeBuilderEntry::TypeAlias(alias) => {
-                    let mutex = type_builder.type_alias(&alias.elem.name);
-                    let alias_builder = mutex.lock().unwrap();
-                    alias_builder.target(alias.elem.r#type.elem.to_owned());
-                }
-            }
-        }
+        type_builder.add_entries(test.type_builder_contents());
 
         type_builder
             .recursive_type_aliases()
